#25 拆出词法部分, 通过基本unittest测试
